requirements.txt
azure-ai-vision
azure-core